deltas are in seconds ;)
Revert "changed delta to ms"

This reverts commit dfdd014fb74d71b3183b830b76cd6b7e8fc63cce.

+++ src/main/java/org/getspout/api/entity/Controller.java
@@ -7,11 +7,5 @@ public abstract class Controller {
 	}
 	
 	public abstract void onAttached();
-	
-	/**
-	 * Called once per tick
-	 * 
-	 * @param dt milliseconds since the last tick
-	 */
-	public abstract void onTick(long dt);
+	public abstract void onTick(float dt);
 }
